requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.10.6
django-crispy-forms==1.6.1
django-pagedown==0.1.1
gunicorn==19.7.1
packaging==16.8
pyparsing==2.2.0
six==1.10.0
whitenoise==3.3.0
psycopg2==2.6.2
